+++ R/rcmd.R
@@ -21,6 +21,8 @@
 rcmd_build_tools <- function(..., env = character(), required = TRUE, quiet = FALSE) {
   env <- c(callr::rcmd_safe_env(), env)
 
+  warn_for_potential_errors()
+
   res <- with_build_tools(
     callr::rcmd_safe(..., env = env, spinner = FALSE, show = FALSE,
       echo = FALSE, block_callback = block_callback(quiet)),
@@ -42,4 +44,19 @@ msg_for_long_paths <- function(output) {
       "for more details: https://github.com/r-lib/remotes/issues/84\n")
   }
 }
+
+warn_for_potential_errors <- function() {
+  if (is_windows() && grepl(" ", R.home()) &&
+      getRversion() <= "3.4.2") {
+    warning(immediate. = TRUE,
+      "\n!!! Building will probably fail!\n",
+      "This version of R has trouble with building packages if\n",
+      "the R HOME directory (currently '", R.home(), "')\n",
+      "has space characters. Possible workarounds include:\n",
+      "- installing R to the C: drive,\n",
+      "- installing it into a path without a space, or\n",
+      "- creating a drive letter for R HOME via the `subst` windows command, and\n",
+      "  starting R from the new drive.\n",
+      "See also https://github.com/r-lib/remotes/issues/98\n")
+  }
 }